prepare to implement direct grids
