+++ math/src/main/java/net/daporkchop/lib/math/arrays/grid/Grid3d.java
@@ -16,8 +16,8 @@
 package net.daporkchop.lib.math.arrays.grid;
 
 import lombok.NonNull;
-import net.daporkchop.lib.math.arrays.grid.impl.DoubleArrayGrid3d;
-import net.daporkchop.lib.math.arrays.grid.impl.IntArrayGrid3d;
+import net.daporkchop.lib.math.arrays.grid.impl.heap.HeapDoubleGrid3d;
+import net.daporkchop.lib.math.arrays.grid.impl.heap.HeapIntGrid3d;
 
 /**
  * @author DaPorkchop_
@@ -31,7 +31,7 @@ public interface Grid3d extends Grid2d {
         if (width * height * depth > arr.length) {
             throw new IllegalArgumentException(String.format("Array length %d too short to be used for grid of %dx%dx%d!", arr.length, width, height, depth));
         } else {
-            return new IntArrayGrid3d(arr, startX, startY, startZ, width, height, depth);
+            return new HeapIntGrid3d(arr, startX, startY, startZ, width, height, depth);
         }
     }
 
@@ -43,7 +43,7 @@ public interface Grid3d extends Grid2d {
         if (width * height * depth > arr.length) {
             throw new IllegalArgumentException(String.format("Array length %d too short to be used for grid of %dx%dx%d!", arr.length, width, height, depth));
         } else {
-            return new DoubleArrayGrid3d(arr, startX, startY, startZ, width, height, depth);
+            return new HeapDoubleGrid3d(arr, startX, startY, startZ, width, height, depth);
         }
     }
